Added a database!
We now have a working database that we can use to get info from and put
info into. Also the “view professor profile” is now pulling from the
database, however we still need to make a method to add to it and we
need to be able to combine data to formulate the proper info for each
field.